对接权限中心

+++ app-mvc/src/main/java/com/datarecord/webapp/fillinatask/bean/RcdJobPersonAssign.java
@@ -1,8 +1,10 @@
 package com.datarecord.webapp.fillinatask.bean;
 
+import java.math.BigInteger;
+
 public class RcdJobPersonAssign {
-    private int user_id;
-    private int  origin_id;
+    private BigInteger user_id;
+    private BigInteger  origin_id;
     private String user_name_cn;
 
     public String getUser_name_cn() {
@@ -13,19 +15,19 @@ public class RcdJobPersonAssign {
         this.user_name_cn = user_name_cn;
     }
 
-    public int getUser_id() {
+    public BigInteger getUser_id() {
         return user_id;
     }
 
-    public void setUser_id(int user_id) {
+    public void setUser_id(BigInteger user_id) {
         this.user_id = user_id;
     }
 
-    public int getOrigin_id() {
+    public BigInteger getOrigin_id() {
         return origin_id;
     }
 
-    public void setOrigin_id(int origin_id) {
+    public void setOrigin_id(BigInteger origin_id) {
         this.origin_id = origin_id;
     }
