+++ app-mvc/src/main/java/com/datarecord/webapp/sys/origin/entity/Origin.java
@@ -1,13 +1,14 @@
 package com.datarecord.webapp.sys.origin.entity;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
 public class Origin {
 
-    private Integer origin_id;
+    private BigInteger origin_id;
     private String origin_name;
-    private Integer parent_origin_id;
+    private BigInteger parent_origin_id;
     private String origin_status;
     private Date create_date;
     private String create_user;
@@ -25,11 +26,11 @@ public class Origin {
     private List<Origin> children;
 
 
-    public Integer getOrigin_id() {
+    public BigInteger getOrigin_id() {
         return origin_id;
     }
 
-    public void setOrigin_id(Integer origin_id) {
+    public void setOrigin_id(BigInteger origin_id) {
         this.origin_id = origin_id;
     }
 
@@ -41,11 +42,11 @@ public class Origin {
         this.origin_name = origin_name;
     }
 
-    public Integer getParent_origin_id() {
+    public BigInteger getParent_origin_id() {
         return parent_origin_id;
     }
 
-    public void setParent_origin_id(Integer parent_origin_id) {
+    public void setParent_origin_id(BigInteger parent_origin_id) {
         this.parent_origin_id = parent_origin_id;
     }
