+++ app-mvc/src/main/java/com/datarecord/webapp/sys/user/service/CustomerUserService.java
@@ -5,22 +5,23 @@ import com.datarecord.webapp.sys.origin.entity.OriginNature;
 import com.datarecord.webapp.sys.user.entity.UserForgetPwdRecord;
 import com.webapp.support.page.PageResult;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Map;
 
 public interface CustomerUserService {
 
-    public PageResult pageCqnyUser(Integer currPage, Integer pageSize, String user_name, String user_type, List<Integer> originList);
+    public PageResult pageCqnyUser(Integer currPage, Integer pageSize, String user_name, String user_type, List<BigInteger> originList);
 
     void selectOriginType(String userId, String originType);
 
-    Map<String, String> getSmsValidateCode(Integer userId, String phone_num);
+    Map<String, String> getSmsValidateCode(BigInteger userId, String phone_num);
 
-    UserForgetPwdRecord getUserForgetPwdRecord(Integer userId);
+    UserForgetPwdRecord getUserForgetPwdRecord(BigInteger userId);
 
-    String updateValidateCode(Integer user_id);
+    String updateValidateCode(BigInteger user_id);
 
-    String newValidateCode(Integer user_id);
+    String newValidateCode(BigInteger user_id);
 
     List<ChinaAreaCode> getAreaData(String parentId);
